Patch for higher resolution of limit_req module.
To combat requests from spam lower rates than one per minute
are required. This patch increases limit_req's resolution
by a factor of 10 and introduces "r/h" rate.

Although resolution is still not high enough for 1 r/h, but it is
close enough to be useful. And it becomes perfect for 3 r/h and
higher rates.

Higher resolution than that requires a bit more work. Switching
to 64-bit ints, to be precise.